@@ -12,49 +12,38 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.2.9",
-    "@angular/cdk": "^7.3.5",
-    "@angular/common": "~7.2.9",
-    "@angular/compiler": "~7.2.9",
-    "@angular/core": "~7.2.9",
-    "@angular/forms": "~7.2.9",
-    "@angular/http": "~7.2.9",
-    "@angular/material": "^7.3.5",
-    "@angular/platform-browser": "~7.2.9",
-    "@angular/platform-browser-dynamic": "~7.2.9",
-    "@angular/router": "~7.2.9",
-    "@kolkov/angular-editor": "^0.13.1",
-    "codemirror": "^5.44.0",
-    "core-js": "^2.6.5",
+    "@angular/animations": "~8.0.3",
+    "@angular/cdk": "^8.0.1",
+    "@angular/common": "~8.0.3",
+    "@angular/compiler": "~8.0.3",
+    "@angular/core": "~8.0.3",
+    "@angular/forms": "~8.0.3",
+    "@angular/http": "~7.2.15",
+    "@angular/material": "^8.0.1",
+    "@angular/platform-browser": "~8.0.3",
+    "@angular/platform-browser-dynamic": "~8.0.3",
+    "@angular/router": "~8.0.3",
+    "@kolkov/angular-editor": "^0.14.6",
+    "codemirror": "^5.48.0",
+    "core-js": "^3.1.4",
     "hammerjs": "^2.0.8",
-    "jquery": "^3.3.1",
     "moment": "^2.24.0",
-    "rxjs": "~6.4.0",
+    "rxjs": "~6.5.2",
     "spinkit": "^1.2.5",
-    "zone.js": "~0.8.29"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.6",
-    "@angular/cli": "~7.3.6",
-    "@angular/compiler-cli": "~7.2.9",
-    "@angular/language-service": "~7.2.9",
-    "@types/codemirror": "^0.0.72",
-    "@types/jasmine": "~3.3.11",
+    "@angular-devkit/build-angular": "~0.800.4",
+    "@angular/cli": "~8.0.4",
+    "@angular/compiler-cli": "~8.0.3",
+    "@angular/language-service": "~8.0.3",
+    "@types/codemirror": "^0.0.75",
+    "@types/jasmine": "~3.3.13",
     "@types/jasminewd2": "~2.0.6",
     "@types/moment": "^2.13.0",
-    "@types/node": "~11.11.3",
-    "@types/socket.io-client": "^1.4.32",
-    "codelyzer": "~5.0.0-beta.1",
-    "jasmine-core": "~3.3.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.0.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.5",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "~5.4.2",
-    "ts-node": "~8.0.3",
-    "tslint": "~5.14.0",
-    "typescript": ">=3.1.1 <3.3"
+    "@types/node": "~12.0.10",
+    "ts-node": "~8.3.0",
+    "tslint": "~5.18.0",
+    "typescript": "~3.4.5"
   }
 }
